Border fix

+++ Assets/Scripts/Infrastructure/Factory/IGameFactory.cs
@@ -7,6 +7,6 @@ namespace Pewpew.Infrastructure.Factory
     {
         GameObject CreatePlayer(GameObject at);
         GameObject CreateLobbyHud();
-        GameObject CreateGameBorder(int borderSize, GameObject at);
+        GameObject CreateGameBorder(float borderSize, GameObject at);
     }
 }
